requirements.txt
aiogram~=3.13.1
yfinance~=0.2.44
python-dotenv~=1.0.1
aiosqlite~=0.20.0
requests~=2.32.3
aiohttp~=3.10.10